% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_double.R
\name{toy_double}
\alias{toy_double}
\title{Title}
\usage{
toy_double(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a numerical vector, doubled
}
\description{
Title
}
\examples{
x <- c(1,2,3)
toy_double(x)
}
